requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt

psycopg2-binary
dj-database-url
python-decouple

daphne
channels
channels-redis

celery
redis

drf-spectacular
django-cors-headers

cryptography
Pillow
numpy
